% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geomtimeline.R
\name{draw_timeline_panel}
\alias{draw_timeline_panel}
\title{draw_timeline_panel}
\usage{
draw_timeline_panel(data, panel_scales, coord)
}
\arguments{
\item{data, panel_scales, coord}{data needed for panel function sent from ggproto instance}
}
\value{
timeline with points
}
\description{
panel function for geom_timeline.  NOT EXPORTED
}
\note{
this is not an exported function, and is not meant to be run outside of calls to geom_timeline(); accordingly, there is no example
}
