% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{Data on earthquakes}
\format{
A data frame with 3 755 rows and 9 variables.
\describe{
  \item{date}{The date of the event as YYYY-MM-DD. (from 1900-01-01 until 2021-04-28)}
  \item{country}{The country where the earthquake effects were observed.}
  \item{region}{Regional boundaries defined by NOAA. Please consult the NOAA website.}
  \item{area}{For US territory:The State, Province or Prefecture where the earthquake effects were observed. Please consult the NOAA website for more information.}
  \item{location_name}{Per NOAA website: "The Country, State, Province or City where the Earthquake occurred (For example enter USA or California or San Francisco). This is only an approximate geographic location.(...)"}
  \item{latitude}{Latitude: 0 to 90 (Northern Hemisphere) -90 to 0 (Southern Hemisphere).}
  \item{longitude}{Longitude: 0 to 180 (Eastern Hemisphere) -180 to 0 (Western Hemisphere).}
  \item{mag}{The value in this column contains the primary earthquake magnitude. (0.0 to 9.9)}
  \item{total_deaths}{Total Number of Deaths from the Earthquake and secondary effects (eg Tsunami)}
}
}
\source{
National Geophysical Data Center / World Data Service (NGDC/WDS): NCEI/WDS Global Significant Earthquake Database. NOAA National Centers for Environmental Information. doi:10.7289/V5TD9V7K .
}
\usage{
data(earthquakes)
}
\description{
The dataset is based on NOAA website.
}
\keyword{datasets}
