% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_earthquakes.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Reads NOAA earthquakes data set from .txt file and convert to clean datafrmae object}
\usage{
eq_clean_data(raw_data_file)
}
\arguments{
\item{raw_data_file}{raw NOAA dataframe .txt file location}
}
\value{
clean dataframe with Date, latitude, longitude conversion.
}
\description{
This function takes raw NOAA data frame and returns a clean data frame-
 1) A date column created by uniting the year, month, day
     and converting it to the Date class
 2) LATITUDE and LONGITUDE columns converted to numeric class
}
\examples{
\dontrun{
 eq_clean_data("signif.txt")
}

}

