% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_earthquakes.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates an HTML label to use in Popup text for Location, Magnitude and Total number of Deaths.}
\usage{
eq_create_label(data_eq)
}
\arguments{
\item{data_eq}{Filtered dataframe of NOAA earthquakes dataset}
}
\value{
Popup text in Html lable for Location, Magnitude and Total number of Deaths
}
\description{
Creates an HTML label that can be used as the annotation text in the leaflet map.
This function put together a character string for each earthquake that will show
the cleaned location , the magnitude (EQ_PRIMARY), and
the total number of deaths (TOTAL_DEATHS),
with boldface labels for each ("Location", "Total deaths", and "Magnitude").
If an earthquake is missing values for any of these, both the label and the
value will be skipped for that element of the tag.
}
\examples{
\dontrun{
 eq_create_label(data_eq)

 data_eq \%>\%
       dplyr::filter(COUNTRY \%in\% COUNTRIES & lubridate::year(DATE) >= 2000) \%>\%
       dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
       eq_map(annot_col = "popup_text")
}

}

