% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_earthquakes.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Converts Location name to Title case}
\usage{
eq_location_clean(eq_dataset_location)
}
\arguments{
\item{eq_dataset_location}{NOAA earthquakes dataset}
}
\value{
Location name with Title case
}
\description{
cleans the LOCATION_NAME column by stripping out the country name (including the colon) and
 converts names to title case (as opposed to all caps). This will be needed later for
 annotating visualizations.
}
\examples{
\dontrun{
 eq_data$LOCATION_NAME <- eq_location_clean(eq_data)
}

}

