% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{A ggplot2 layer: \code{geom_timeline_label()} function}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mappings created by aes}

\item{data}{is the data frame that contains the earthquake's data}

\item{stat}{layer's statistical transformation}

\item{position}{position adjustment function}

\item{na.rm}{remove the NA values from the data frame}

\item{show.legend}{layer's legend}

\item{inherit.aes}{will indicate the default aesthetics overridng}

\item{...}{layer's other arguments}
}
\value{
a plot timeline with earthquakes of a given country or list of countries.
}
\description{
Construct a regular function/layer to plot a timeline with the earthquakes
of a given country that is added to a plot created with ggplot() function. This function is
for adding annotations to the earthquake data. This geom adds a vertical line to each data
point with a text annotation (e.g. the location of the earthquake) attached to each line.

The \code{geom_timeline_label()} function is a wrapper to the layer function for the 'GeomTimelineLabel' class.
}
