% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timelinelabel}
\alias{geom_timelinelabel}
\title{Annotations to the earthquake data}
\usage{
geom_timelinelabel(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings}

\item{data}{The dataset}

\item{stat}{stat layer}

\item{position}{Position adjustment}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{show.legend}{legend to display}

\item{inherit.aes}{false: default aesthetics}

\item{...}{Other arguments passed on to [layer()].}
}
\value{
a geom
}
\description{
This geom adds a vertical line to each data point with a text annotation (e.g. the location of the earthquake) attached to each line.
}
\details{
Use \code{earthquakes} dataframe from this package or download new dataset from NOAA site.
}
\examples{
\donttest{
library(ggplot2)
a<-ggplot(earthquakes,aes(date,location_name,col=total_deaths,size=mag))+
geom_timelinelabel()
}
}
