% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes_plotting.r
\docType{data}
\name{GeomTime}
\alias{GeomTime}
\title{Creating a Geom -> GeomTime}
\format{
An object of class \code{GeomTime} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTime
}
\value{
The function returns a Geom to be used by geom_timeline to render
the images into the plot.
}
\description{
This Geom returns a ggproto object for our geom to inherit from.
}
\keyword{datasets}
