% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomTimeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Class GeomTimeline}
\format{
An object of class \code{GeomTimeline} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\description{
Construct a new class 'GeomTimeline' which is a 'ggproto' object
for plotting a time line of earthquakes.#'
}
\keyword{datasets}
