% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomTimelineLabel.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Class GeomTimelineLabel}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{GeomText}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomTimelineLabel
}
\description{
Construct a new class 'GeomTimelineLabel' which is a 'ggproto' object
for plotting a time line of earthquakes.#'
}
\keyword{datasets}
