% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Loads the raw NOAA significant earthquake data.}
\usage{
load_data()
}
\value{
a dataframe containing the raw NOAA significant earthquake data.
}
\description{
Loads the raw NOAA significant earthquake data.
}
