% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{read_naoo_file}
\alias{read_naoo_file}
\title{Function to read NAOO file.}
\usage{
read_naoo_file(filename)
}
\arguments{
\item{filename}{A filename as input string.}
}
\value{
File data is read into a dataframe object.
}
\description{
This function uses readr and dplyr packages.
1.   Read file
     Takes a TAB delimited file as an input filename parameter.
     If the file exists, this function reads it into dataframe object.
2.   Error handling
     If the file does not exist in the working directory, this function stops with an output message that the file does not exist.
3.   Suppressions
     It suppresses messages and progress bar while reading the file.
}
\examples{
\dontrun{
read_naoo_file("signif.txt")
read_naoo_file(filename="signif.txt")
}

}

