% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean.R
\name{RemoveBefore}
\alias{RemoveBefore}
\title{RemoveBefore}
\usage{
RemoveBefore(str, rmv)
}
\arguments{
\item{str}{string from which to remove characters.  Works with lists of strings}

\item{rmv}{pattern marking where to cut}
}
\value{
string to return.  if is.na, returns 'ocrean', indicating that the quake occurred at sea (not an exported function for a reason)
}
\description{
removes all characters from string before pattern rmv
}
\note{
NOT EXPORTED
}
