% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_earthquakes.R
\name{them_timeline}
\alias{them_timeline}
\title{Timeline Them to represent earthquakes on yeary timeframe}
\usage{
them_timeline()
}
\value{
Plot them of White background with X aes as timeline
}
\description{
Timeline Them to represent earthquakes on yeary timeframe
}
\examples{
\dontrun{
 them_timeline()
}

}

