% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes_plotting.r
\name{theme_time}
\alias{theme_time}
\title{Customize the plot theme}
\usage{
theme_time()
}
\description{
Plot theme is modified from ggplot2 package.
}
\details{
The parameters used to customize the theme can be found in the
url : https://ggplot2.tidyverse.org/reference/theme.html
}
