% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean.R
\name{ymd_bc}
\alias{ymd_bc}
\title{ymd_bc}
\usage{
ymd_bc(date, ...)
}
\arguments{
\item{date}{a date in string format or a list of dates in string format}

\item{...}{extra variables that may work wtih the lubridate::date function.}
}
\description{
given a date in string format, returns a date in date format.  Works with BC.  Can handle lists
}
\examples{
  test_data <- c('-2000 01 01', '500 05 30','1990 07 04')
  Earthquakes:::ymd_bc(test_data)
}
