% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes-data.R
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{NOAA Earthquakes data example}
\format{An object of class \code{"tbl_df tbl data.frame"}.}
\source{
\href{https://www.ngdc.noaa.gov/nndc/struts/results?type_0=Exact&query_0=$ID&t=101650&s=13&d=189&dfn=signif.txt}{Database as file}
}
\usage{
data(earthquakes)
}
\description{
This data set includes details about earthquakes in Russia, China and USA from NOAA Eathquakes database.
}
\references{
National Geophysical Data Center / World Data Service (NGDC/WDS): Significant Earthquake Database.
National Geophysical Data Center, NOAA. doi:10.7289/V5TD9V7K
(\href{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}{NOAA})
}
\keyword{datasets}
