% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Print "Earthquakes Clean Data"}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{The raw data loaded into R environment after reading in from https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\value{
This function returns a tibble with a date field, numeric latitude and longitude values and a city name.
}
\description{
This function reads a raw dataset of significant earthquake data provide by NOAA (National Centers for Environmental Information).
}
\note{
Earthquakes that occurred BCE (Before Current Epoch) have been removed from the dataset because negative year values are not handled well.
}
\examples{
\dontrun{
eq_clean_data(raw)
}

}
