% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Read the earthquake data from file and tidy it up}
\usage{
eq_clean_data(filename)
}
\arguments{
\item{filename}{character vector with the file path containing the earthquake data}
}
\value{
data table that contains the cleaned earthquake data
}
\description{
Read the earthquake data from file and tidy it up
}
\examples{
\dontrun{
eqdata <- eq_clean_data ("data/signif.txt")
}

}
