% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{quake.txt data to be cleaned}
}
\description{
given quake data read from quakes.txt, adds DATE column
}
