% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans the raw NOAA significant earthquake data.}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{the raw NOAA significant earthquake dataset}
}
\value{
the cleaned up data set
}
\description{
This converts latitude and longitude into numeric variables,
converts various other variables to numeric type, and creates
a new variable called DATE, which is of type Date and contains
the combination of the YEAR, MONTH and DAY variables.
}
\details{
Careful: NA values in MONTH and DAY are replaces with 1, e.g.
YEAR = 2019, MONTH = NA, DAY = NA will result in DATE = 2019-01-01.
Negative years return in NA in the DATE variable, i.e. the historic
earthquakes are not correctly handled.

Additionally, the LOCATION_NAME column is cleaned up using the
eq_location_clean function.
}
\examples{
\dontrun{
data <- readr::read_delim("noaa_signif_earthquake.txt", delim = "\\t")
clean_data <- eq_clean_data(data)
}

}
