% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleaning Earthquake data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{Raw earthquake data in the dataframe in format of NOAA earthquakes database.}
}
\value{
Data frame with cleaned up data
}
\description{
Under the cleaning we understand the following:
new DATE column added;
LOCATION cleaned up;
lines with no LONGITUDE or LATITUDE removed;
lines with no Richter magnitude removed;
data with YEAR B.C. removed.
}
\examples{

filename <- system.file("extdata", "earthquakes.tsv.gz", package = "earthquakes")
raw_data <- read_delim(filename, delim = "\\t")
data <- eq_clean_data(raw_data)

}
