% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidying.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA Earthquake Data}
\usage{
eq_clean_data(raw_data = rlang::.data)
}
\arguments{
\item{raw_data}{A data frame containing the raw NOAA earthquake data with the following required columns: Year, Mo, Dy, Latitude, Longitude, Location.Name, Hr, Mn, and Sec.}
}
\value{
A cleaned data frame with additional columns and removed unnecessary columns.
}
\description{
The function cleans the NOAA earthquake dataset by creating a date column, converting latitude and longitude to numeric, extracting country names, and cleaning location names.
}
\examples{
\dontrun{
raw_data <- read.delim(system.file("extdata", "earthquakes.tsv",
  package = "earthquakes"
), header = TRUE, sep = "\t")
cleaned_data <- eq_clean_data(raw_data)
head(cleaned_data)
}
}
\keyword{internal}
