% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Function to clean National Oceanic and Atmospheric Administration's (NOAA) file.}
\usage{
eq_clean_data(filename)
}
\arguments{
\item{filename}{A filename as input string.}
}
\value{
File data is read into a dataframe object with clean data in DATE, LATITUDE, LONGITUDE and LOCATION_NAME fields.
}
\description{
This function uses dplyr, lubridate, tidyr,  and  tools packages.
1.   Read file
     Takes a TAB delimited file as an input filename parameter.
2.   Create DATE field
     Filters out NA or 0 YEAR field data, NA MONTH field data, NA DAY field data.
     Combines YEAR, MONTH and DAY fields from the file and converts it into a date field and drops NA dates.
3.   LATITUDE and LONGITUDE fields data
     Mutates LATITUDE and LONGITUDE fields data to numeric.
4.   LOCATION_NAME field data
     Cleans LOCATION_NAME field data by stripping of the country name, extra whitespaces.
     Cleaned LOCATION_NAME field data has city name or locaton names in the country.
     Converts cleaned LOCATION_NAME field data to toTitleCase.
}
\examples{
\dontrun{
eq_clean_data("signif.txt")
eq_clean_data(filename="signif.txt")
}

}

