% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleaning a NOAA (earthquake) file}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{- A delimited file that is read in by the 'read_delim' function}
}
\value{
A cleaned NOAA data frame
}
\description{
The 'eq_clean_data' function returns a cleaned NOAA data frame
}
\details{
The eq_clean_data function correct the data type of the columns YEAR, MONTH, DAY, LATITUDE, LONGITUDE, DEATHS and EQ_PRIMARY.
         The columns YEAR, MONTH and DAY are gathered to form a date (an as.Date object). The missing dates are filtered out. The
         magnitude of the earthquakes are rounded up to 2, 4, 6, 8 and 10. The locations in the column LOCATION_NAME is extracted and
         adjusted to title case. The column COUNTRY is alsp adjusted to title case with the eq_location_clean fucntion.
}
\note{
error - When the column names do not include YEAR, MONTH, DAY, LATITUDE, LONGITUDE, DEATHS, EQ_PRIMARY, LOCATION_NAME
              and COUNTRY, an error will be given.
}
\examples{
\dontrun{
file = "inst/extdata/earthquakes.tsv.gz"
file <- readr::read_delim(file, delim = "\\t")
file_cleaned <- eq_clean_data(file)
}


}
