% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes_cleaning.r
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Data from the NOAA data frame.}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{Earthquake dataset.}
}
\value{
The function returns the clean dataset and calls the next function
eq_location_clean().
}
\description{
This function cleans the data, mainly the following columns.
YEAR, MONTH, DAY, LATITUDE, LONGITUDE, MAGNITUDE, COUNTRY and DEATHS
}
\examples{
\dontrun{
 readr::read_delim("earthquakes-2021-09-27_15-28-44_+0200.tsv") \%>\%
 eq_clean_data()
}

}
