% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean dataset}
\usage{
eq_clean_data(x)
}
\arguments{
\item{x}{A messy datasets with columns as described}
}
\value{
This function returns a (hopefully cleaner) dataset
}
\description{
This function tidies the dataset up by creating a DATE column from the YEAR,
MONTH and DAY columns (with date format) and converting the columns LONGITUDE
and LATITUDE into numeric.
}
\examples{
\dontrun{
eq_clean_data(messy_data)
}

}
