% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean data frame}
\usage{
eq_clean_data(rawdata)
}
\arguments{
\item{rawdata}{a data frame obtained from NOAA website.}
}
\value{
A date column created by uniting the year, month, day and converting it to the Date class. Plus, LATITUDE and LONGITUDE columns converted to numeric class.
}
\description{
Takes raw NOAA data frame and returns a clean data frame.
}
\examples{
\dontrun{
raw_data <- readr::read_delim("earthquakes.tsv", "\t")
eq_clean_data(raw_data)
}

}
