% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans the raw NOAA earthquake dataset.}
\usage{
eq_clean_data(eq_raw)
}
\arguments{
\item{eq_raw}{the raw NOAA earthquake dataset as a data frame that contains location names written in upper case.}
}
\value{
returns a clean data frame of the NOAA earthquake dataset.
}
\description{
Takes raw NOAA data frame and returns a clean data frame. The clean data frame should have the following:
1. A date column created by uniting the year, month, day and converting it to the Date class.
2. LATITUDE and LONGITUDE columns converted to numeric class.
3. In addition, a LOCATION_NAME column by stripping out the country name (including the colon)
and converts names to title case (as opposed to all caps). This will be needed later for annotating visualizations.
}
\examples{

}
