% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqmaps.R
\name{eq_create_label}
\alias{eq_create_label}
\title{This method returns a popup-text for each in the data frame with location name, magnitude and
deaths.}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{The data frame for the popup-texts.}
}
\value{
list of popup-texts.
}
\description{
This method returns a popup-text for each in the data frame with location name, magnitude and
deaths.
}
\examples{
\dontrun{
 f <- system.file("extdata", "signif.txt", package="earthquakes")
 a <- eq_clean_data(file = f)
 b <- dplyr::filter(a, COUNTRY == "MEXICO" && lubridate::year(DATE) >= 2000)
 c <- dplyr::mutate(b, popup_text = eq_create_label(.))
 eq_map(c, annot_col = "popup_text")
}
}
