% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqmaps.R
\name{eq_create_label}
\alias{eq_create_label}
\title{This method returns a popup-text for each in the data frame with location name, magnitude and
deaths.}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{The data frame for the popup-texts.}
}
\value{
list of popup-texts.
}
\description{
This method returns a popup-text for each in the data frame with location name, magnitude and
deaths.
}
\examples{
\dontrun{
 eq_clean_data("data/signif.dat") \%>\%
 dplyr::filter(COUNTRY == "MEXICO" && lubridate::year(DATE) >= 2000) \%>\%
 dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
 eq_map(annot_col = "popup_text")
}
}
