% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Returns an annotation vector based on a NOAA earthquake dataset.
The annotation contains information about location, magnitude
and total deaths. If any information is missing, it is skipped.}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{the cleaned NOAA earthquake dataset}
}
\value{
a character vector containing the annotations labels
}
\description{
Returns an annotation vector based on a NOAA earthquake dataset.
The annotation contains information about location, magnitude
and total deaths. If any information is missing, it is skipped.
}
\examples{
\dontrun{
eq_create_label(data)
}

}
