% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Make a HTML description of the earthquakes}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Dataframe with earthquake data. Structure and columns names must be the same as in NOAA earthquake database.}
}
\value{
Character vector of earthquakes description. Description consist of 3 lines: Location, Magnitude, deaths.
If magnitude or deaths is absent, location only provided.
}
\description{
Function eq_create_label() takes the dataset as an argument and creates an HTML label that can be used
as the annotation text in the leaflet map. This function put together a character string for each earthquake that
will show the cleaned location (as cleaned by the eq_location_clean() function descibed above),
the magnitude (EQ_PRIMARY), and the total number of deaths (TOTAL_DEATHS),
with boldface labels for each ("Location", "Total deaths", and "Magnitude").
If an earthquake is missing values for any of these, both the label and the value are skipped for
that element of the tag.
}
\examples{

eq_create_label(earthquakes)

}
