% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Make a HTML description of the earthquakes}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Dataframe with earthquake data. Structure and columns names must be the same as in NOAA earthquake database.}
}
\value{
Character vector of earthquakes description. Description consist of 3 lines: Location, Magnitude, deaths.
If magnitude or deaths is absent, location only provided.
}
\description{
Make a HTML description of the earthquakes
}
\examples{

eq_create_label(earthquakes)

}
