% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Function to create label for loation, magnitude and deaths.}
\usage{
eq_create_label(eq_data)
}
\arguments{
\item{eq_data}{Tidy earthquake data}
}
\value{
HTML label with location, magnitude and deaths
}
\description{
This function uses shiny package.
}
\examples{
\dontrun{
eq_create_label(data=eq_data)
}

}

