% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Labelling the pop-up of the geographical map with earthquakes}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{- file name}
}
\value{
A pop-up with information about the location, magnitude and the number of deaths, when hovering over an earthquake.

#@example
#\dontrun{
#x <- readr::read_delim("data/earthquakes.tsv.gz", delim = "\t") %>%
#  eq_clean_data() %>%
#  dplyr::filter(COUNTRY == "Mexico" & lubridate::year(DATE) >= 2000) %>%
#  dplyr::mutate(popup_text = eq_create_label(.)) %>%
#  eq_map(annot_col = "popupText")
#x
#}
}
\description{
Information about the earthquakes like locations, magnitude and the number of deaths are given in a pop-up.
}
\details{
The 'eq_create_label' function indicates 3 information about the earthquakes in the pop-up, when hovering over an earthquake.
         A line about the locations of the earthquake, a line about the magnitude of the earthquakes and a line about the number of
         deaths.
}
\note{
error - An error will be given when there are no column named as "locations", "EQ_PRIMARY" and "DEATHS".
}
