% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes_mapping.r
\name{eq_create_label}
\alias{eq_create_label}
\title{Introduce Labels in a Leaflet Map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{The dataset to plot.}
}
\description{
This function plots a leaflet map and labels it.
}
\examples{
\dontrun{
readr::read_delim("earthquakes-2021-09-27_15-28-44_+0200.tsv") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY \%in\% c("MEXICO") & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\% eq_map(annot_col = "popup_text")
}

}
