% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create additional column combining information from the dataset}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A dataset with columns as described above}
}
\value{
This function returns the original dataset plus an additional column as
described above.
}
\description{
This function takes in input a dataset with LOCATION, EQ_PRIMARY and TOTAL_DEATHS columns.
Then creates an additional html formatted column combining the information from these three
column.
}
\examples{
\dontrun{
mutate(data, label = eq_create_label(.))
}

}
