% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Map HTML labels}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{a data frame. Use earthquakes or download dataset from NOAA website.}
}
\value{
An HTML label to be used as annotation text in the leaflet map.
}
\description{
Takes the dataset as an argument and creates an HTML label that can be used as the annotation text in the leaflet map.
}
\examples{
\dontrun{
earthquakes \%>\%
  dplyr::filter(country == "MEXICO" & lubridate::year(date) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}

}
