% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates popup text for markers.}
\usage{
eq_create_label(eq_data_clean = NULL)
}
\arguments{
\item{eq_data_clean}{The clean earthquake data in a tbl_df object.}
}
\value{
This function returns a character vector containing popup text to be used in a leaflet visualization.
}
\description{
This function generates HTML formatted text to be used in popups for map markers.
}
\examples{
\dontrun{
filename <- system.file("extdata", "signif.txt", package = "earthquakes")
library(readr)
eq_data_raw <- readr::read_delim(file = filename, delim = "\t")
eq_data <- eq_clean_data(eq_raw = eq_data_raw)
library(dplyr)
library(lubridate)
library(leaflet)
eq_data \%>\%
    filter(COUNTRY \%in\% "ITALY" &
               lubridate::year(DATE) >= 2000) -> eq_data_sample

eq_data_sample \%>\%
    eq_clean_data() \%>\%
    filter(COUNTRY \%in\% "ITALY" & lubridate::year(DATE) >= 2000) \%>\%
    mutate(popup_text = eq_create_label(.)) \%>\%
    eq_map(annot_col = "popup_text")
}

}
