% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{quake data--designed for builtin dataset}
}
\value{
quake data with cleaned location names
}
\description{
given strings of location name, removes everything before the ':'
}
