% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans up location names}
\usage{
eq_location_clean(location)
}
\arguments{
\item{location}{a character vector containing the raw location names}
}
\value{
the cleaned up location names character vector
}
\description{
The NOAA significant earthquake dataset contains location names
that are all uppercase and consist of country, optionally region
and location, separated by a colon.
}
\details{
This function strips country and region and converts the location
to title case.
}
\examples{
\dontrun{
data <- readr::read_delim("noaa_signif_earthquake.txt", delim = "\\t")
clean_data <- eq_location_clean(data)
}

}
