% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Earthquake location from NOAA Earthquake database}
\usage{
eq_location_clean(raw_earthquake)
}
\arguments{
\item{raw_earthquake}{Raw earthquake data in the dataframe in format of NOAA earthquakes database.}
}
\value{
data frame with cleaned up LOCATION_NAME field
}
\description{
Function eq_location_clean() cleans the LOCATION_NAME column by stripping out the country name
(including the colon) and converts names to title case (as opposed to all caps).
}
\examples{
filename <- system.file("extdata", "earthquakes.tsv.gz", package = "earthquakes")
raw_data <- read_delim(filename, delim = "\\t")
data <- eq_location_clean(raw_data)

}
