% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Function to clean LOCATION_NAME field from National Oceanic and Atmospheric Administration's (NOAA) file.}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{Data as dataframe object.}
}
\value{
Dataframe object with clean LOCATION_NAME fields data.
}
\description{
This function uses dplyr and tools packages.
Cleans LOCATION_NAME field data by stripping of the country name, extra whitespaces.
Cleaned LOCATION_NAME field data has city name or locaton names in the country.
Converts cleaned LOCATION_NAME field data to toTitleCase.
}
\examples{
\dontrun{
eq_location_clean(data=eq_clean)
}

}

