% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes_cleaning.r
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Format LOCATION_NAME}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{dataframe containing a column with the name "LOCATION_NAME".}
}
\value{
This function returns location name with the title case format.
}
\description{
This function removes the country and the ":" symbol from the location name.
}
\examples{
\dontrun{
 data <- data.frame(LOCATION_NAME = "SPAIN:  BARCELONA")
 eq_location_clean(data)
}

}
