% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean the LOCATION_NAME column}
\usage{
eq_location_clean(rawdata)
}
\arguments{
\item{rawdata}{a data frame obtained from NOAA website. Alternatively, this package already contains a dataset from NOAA. Check raw_data for details.}
}
\value{
a data frame with a cleaned up version of the Location_Name column.
}
\description{
cleans the LOCATION_NAME column by stripping out the country name (including the colon) and converts names to title case (as opposed to all caps).
}
\details{
This function should be applied to the raw data to produce a cleaned up version of the LOCATION_NAME column.
}
\examples{
\dontrun{
raw_data <- readr::read_delim("earthquakes.tsv", "\t") \%>\%
eq_location_clean()
}

}
