% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans the location name of the raw NOAA earthquake dataset.}
\usage{
eq_location_clean(eq_datf)
}
\arguments{
\item{eq_datf}{the raw NOAA earthquake dataset as a data frame that contains location names written in upper case.}
}
\value{
returns a clean data frame of the NOAA earthquake dataset.
}
\description{
Cleans the (\code{LOCATION_NAME}) column by stripping out the country name (including the colon)
and converts names to title case (as opposed to all caps). This will be needed later for annotating
visualizations. This function should be applied to the raw data to produce a cleaned up version of
the (\code{LOCATION_NAME}) column.
}
\examples{
\dontrun{
filename <- system.file("extdata", "signif.txt", package = "earthquakes")
library(readr)
eq_data_raw <- readr::read_delim(file = filename, delim = "\t")
eq_data <- eq_location_clean(eq_datf = eq_data_raw)
head(eq_data_raw$LOCATION_NAME)
head(eq_data$LOCATION_NAME)
}

}
