% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqmaps.R
\name{eq_map}
\alias{eq_map}
\title{This method uses leaflet to plot latitude and longitude for earthquakes on the map.
The map is interactive with markers that are clickable for showing a text.}
\usage{
eq_map(df, annot_col)
}
\arguments{
\item{df}{The data frame with earthquake data.}

\item{annot_col}{The column name with the text displayed for the markers.}
}
\value{
The leaflet as htmlwidget with the interactive map
}
\description{
This method uses leaflet to plot latitude and longitude for earthquakes on the map.
The map is interactive with markers that are clickable for showing a text.
}
\examples{
\dontrun{
 readr::read_delim("data/signif.dat", delim = "\\t") \%>\%
 eq_clean_data() \%>\%
 dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
 eq_map(annot_col = "DATE")
}
}
