% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Creates a leaflet map of the NOAA earthquake data
and annotates the earthquakes using the annot_col
column in the data.}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{the cleaned NOAA earthquake dataset}

\item{annot_col}{the column to use for annotation labels}
}
\value{
the leaflet map
}
\description{
Creates a leaflet map of the NOAA earthquake data
and annotates the earthquakes using the annot_col
column in the data.
}
\examples{
\dontrun{
eq_map(data, annot_col = "LOCATION_NAME")
}

}
