% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{eq_map}
\alias{eq_map}
\title{Make an Earthquake interactive map}
\usage{
eq_map(data, annot_col = "EQ_PRIMARY")
}
\arguments{
\item{data}{Dataframe with earthquake data. Structure and columns names must be the same as in NOAA earthquake database.}

\item{annot_col}{Name of the column in the data frame used for popup annotation.}
}
\value{
leaflet map object
}
\description{
Make an Earthquake interactive map
}
\examples{
filename <- system.file("extdata", "earthquakes.tsv.gz", package = "earthquakes")
readr::read_delim(filename, delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")

}
