% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_map}
\alias{eq_map}
\title{Function plots map based on data and annot_col.}
\usage{
eq_map(eq_data, annot_col)
}
\arguments{
\item{eq_data}{Tidy earthquake data}

\item{annot_col}{Column name from data used for the annotation in the pop-up}
}
\value{
ggplot2 plot
}
\description{
This function uses leaflet package.
}
\examples{
\dontrun{
eq_map(annot_col = "DATE")
eq_map(annot_col = "popup_text")
}

}

