% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Mapping the earthquakes on a geographical map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{- A file name}

\item{annot_col}{- The parameter that is annotated in a pop-up.}
}
\value{
A geographical map with earthquakes (circles) of which the size indicates the magnitude of the earthquakes.
}
\description{
The 'eq_map' function maps the earthquakes on a geograpical map.
}
\details{
The 'eq_map' function maps the earthquakes (as circles) on a geograpical map, given the longitude and latitude of
         the earthquake events. The magnitude of the earthquakes is shown by the radius of the circles. When hovering over
         an earthquake events a pop-up shows some information about the earthquake.
}
\examples{
\dontrun{
library(magrittr)
x <- readr::read_delim("data/earthquakes.tsv.gz", delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "Mexico",
   DATE >= "2000-01-01") \%>\%
  eq_map(annot_col = "DATE")
x
}

}
