% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes_mapping.r
\name{eq_map}
\alias{eq_map}
\title{Plot Earthquakes in a Map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{The dataset to plot.}

\item{annot_col}{Data representing the markers that will be used in the popup.}
}
\value{
This function returns a leaflet map with the earthquakes positioned in each location.
}
\description{
This function takes a dataset, with latitude and longitude columns, and displays
its earthquakes in a leaflet map. The data is presented as circles with popup info
and the size is proportional to the magnitude registered.
}
\examples{
 \dontrun{
readr::read_delim("earthquakes-2021-09-27_15-28-44_+0200.tsv") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY \%in\% c("MEXICO") & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")
}

}
