% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_map}
\alias{eq_map}
\title{Create leaflet map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A dataset with geographic information and a column with annotations}

\item{annot_col}{A column with annotations in the dataset}
}
\value{
This function returns a leaflet interactive map
}
\description{
This function takes in input a dataset with longitude and latitude columns
and a column to extract labels from. Then creates an interactive leaflet map
with markers annotated by the given column.
}
\examples{
\dontrun{
eq_map(geographic_data)
}

}
