% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Earthquakes Data in an Interactive Map.}
\usage{
eq_map(eq_data_clean = NULL, annot_col = "DATE")
}
\arguments{
\item{eq_data_clean}{the clean earthquake data in a tbl_df object.}

\item{annot_col}{a column name in the tbl_df object to be used for annotation.}
}
\value{
an interactive leaflet map.
}
\description{
The function  \code{eq_map()} that takes an argument data containing
the filtered data frame with earthquakes to visualize. The function maps the
epicenters (\code{LATITUDE/LONGITUDE}) and annotates each point with in pop up window
containing annotation data stored in a column of the data frame.
The user should be able to choose which column is used for the annotation in
the pop-up with a function argument named \code{annot_col}. Each earthquake
should be shown with a circle, and the radius of the circle should be
proportional to the earthquake's magnitude (\code{EQ_PRIMARY}).

The Earthquakes will be mapped centered with their latitude and
longitude "epicenter". The epicenter is annotated based on an \code{annot_col}
which the user can specify. In addition, if the user specifies \code{"popup_text"}
then a call to eq_create_label generates the appropriate text for the popup.
}
\note{
If an invalid column name is provided, the function provides a warning
and uses the LOCATION_NAME column as the annotation column.
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes_data.txt.zip",package="capstone")
eq_location_clean(eq_data_clean_data(eq_data_read(filename))) \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(datetime) >= 1980) \%>\%
eq_map(annot_col = "datetime")
}

}
