% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Print "Earthquakes Timeline Visualization"}
\usage{
geom_timeline(data, start_date, end_date, country)
}
\arguments{
\item{data}{The data has been cleaned using eq_clean_data function}

\item{start_date}{The first date to return data for viz}

\item{end_date}{The last date to return data for viz}

\item{country}{Provide country name like ITALY, NEW ZEALAND}
}
\value{
This function returns a ggplot timeline visualization with DATE on the x-axis, earthquake count on the y-axis where size of point indicates magnitude(EQ_MAG_ML) and color signifies Number of Deaths due to earthquake
}
\description{
This function reads a clean dataset of significant earthquake data provide by NOAA (National Centers for Environmental Information).
}
\note{
Earthquakes that occurred BCE (Before Current Epoch) have been removed from the dataset because negative year values are not handled well.
}
\examples{
\dontrun{
geom_timeline(data, "2000-01-01", "2018-01-01", "NEW ZEALAND")
}

}
