% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqgeoms.R
\name{geom_timeline}
\alias{geom_timeline}
\title{This method creates the required layer containing the timeline geom for earthquake data on.
The color aesthetic shows the number of deaths.
The size aesthetic shows the Richter scale of the earthquakes.}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes or aes_.}

\item{data}{The data to be displayed in this layer, data.frame}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment as a string or function.}

\item{na.rm}{either TRUE or FALSe for for removing missing values with or without warning.}

\item{show.legend}{TRUE or FALSE to show the legend or notlogical.}

\item{inherit.aes}{TRUE or FALSE for inheriting aies}

\item{...}{Additional parameters}
}
\value{
A ggplot2 layer containing the timeline geom for earthquake data.
}
\description{
This method creates the required layer containing the timeline geom for earthquake data on.
The color aesthetic shows the number of deaths.
The size aesthetic shows the Richter scale of the earthquakes.
}
\examples{
\dontrun{
 p <- readr::read_delim("inst/extdata/signif.dat", delim = "\\t") \%>\%
     eq_clean_data() \%>\% eq_location_clean() \%>\%
     dplyr::filter(YEAR >= 2000, COUNTRY \%in\% "USA") \%>\%
     ggplot2::ggplot() +
     geom_timeline(ggplot2::aes(x = DATE, colour = DEATHS,
       size = EQ_MAG_ML)) +
     ggplot2::labs(x = "DATE", color = "Deaths", size = "Richter scale")
 gt <- ggplot2::ggplot_gtable(ggplot2::ggplot_build(p))
 grid::grid.draw(gt)
}
}
