% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Make a timeline geom}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{ggplot parameter}

\item{data}{Earthquake data in the dataframe in format of NOAA earthquakes database.}

\item{stat}{ggplot parameter}

\item{position}{ggplot parameter}

\item{na.rm}{ggplot parameter}

\item{show.legend}{ggplot parameter}

\item{inherit.aes}{ggplot parameter}

\item{...}{other ggplot parameter}

\item{xmindate}{aes parameters: Minimal date of the considering time frame, Date}

\item{xmaxdate}{aes parameters: Maximum date of the considering time frame, Date}

\item{event_date}{aes parameters: date of earthquake, usually name of the column in data (Date)}

\item{richter}{aes parameters: Richter`s magnitude of earthquakes, usually name of the column in data (numeric)}

\item{death}{aes parameters: deaths produced by earthquakes, usually name of the column in data (numeric)}

\item{colour}{aes parameters: main colour of earthquakes marks}

\item{scale}{aes parameters: scale of earthquakes marks}

\item{transparency}{aes parameters: transparency of earthquakes marks}

\item{layers}{aes parameters: groping of earthquakes marks by some parameters, e.g. COUNTRY}
}
\value{
timeline geom (ggplot2 object)
}
\description{
Make a timeline geom
}
\examples{

ggplot2::ggplot() +
  theme_void() +
  geom_timeline(data = earthquakes, aes(event_date = DATE,
                                  xmindate = 2000,
                                  xmaxdate = 2010,
                                  richter = EQ_PRIMARY,
                                  death = DEATHS,
                                  transparency = 0.3,
                                  colour = "red",
                                  scale = 0.5,
                                  layers = COUNTRY))+
  theme(legend.position="none")

}
