% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{geom_timeLine}
\alias{geom_timeLine}
\title{Geom to plot earthquakes by date and magnitude}
\usage{
geom_timeLine(data = NULL, mapping = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, xmin = NULL, xmax = NULL, ...)
}
\arguments{
\item{data}{Tidy earthquake data, as dataframe}

\item{mapping}{Mapping asethetics}

\item{stat}{Value for Statistical transformation, as string}

\item{position}{Position adjustment for the layer, as string}

\item{na.rm}{Remove missing values}

\item{show.legend}{Weather to show legend or not, as NA/TRUE/FALSE/}

\item{inherit.aes}{Weather to show legend or not, TRUE/FALSE/}

\item{xmin}{Minimum Date to plot}

\item{xmax}{Maximum Date to plot}

\item{...}{Other asethetics/parameters for plotting}
}
\value{
Returns ggplot2 labels
}
\description{
This Geom uses ggplot2 package.
}

