% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Geom to plot earthquakes by date and magnitude}
\usage{
geom_timeline(data, xmindate, xmaxdate, country)
}
\arguments{
\item{data}{Tidy earthquake data, as dataframe}

\item{xmindate}{Minimum Date to plot}

\item{xmaxdate}{Maximum Date to plot}

\item{country}{Character vector of Country/Countries to plot}
}
\value{
Returns ggplot
}
\description{
This Geom uses ggplot2 package.
}
\examples{
\dontrun{
geom_timeline(data = eq_clean_data('signif.txt'),
              xmindate = "2001-01-01",
              xmaxdate = "2017-12-31",
              country = c("MEXICO"))
}

}

