% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Visualising a timeline of the earthquakes}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{- File name}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{x}{- The parameter in the X-axis: the time (as.Date object)}

\item{y}{-  The parameter in the Y-axis: the country names (character)}

\item{size}{- The parameter that represents the magnitude of the earthquake, indicating the size of the circles (=earthquake event): Richter scale value (numeric)}

\item{fill}{- The parameter that represents the deaths resulted from the earthquake, colouring the circles (=earthquake event)}
}
\value{
A geom layer with timeline, each timeline shows a country with its earthquake events in circles, the magnitude of an earthquake (size of
        circles) and the number of deaths (the colouring of the circles).
}
\description{
The geom_timeline visualises earthquakes events over time and per country in magnitude and number of deaths.
}
\details{
The geom_timeline function works together with ggplot2. This function visualises the earthquakes events over time (x aesthetics)
         and per country (y aesthetics) in magnitude (size parameter) and number of deaths (fill parameter.
}
\note{
error - An error will appear when a parameter is not given.
}
\examples{
\dontrun{
ggplot() +
  geom_timeline(eq_clean_data(file) \%>\%
       filter(COUNTRY == c("China", "Usa", "Japan"),
             DATE >= "1999-01-01",
             DATE <= "2012-12-31"),
             aes(x = DATE, y = COUNTRY, size = richterScaleValue, fill = DEATHS)) +
  theme(legend.position = "bottom")
}

}
