% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{Geom_Timeline}
\alias{Geom_Timeline}
\title{Visualising a timeline of the earthquakes}
\format{An object of class \code{Geom_Timeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
Geom_Timeline
}
\arguments{
\item{data}{- File name}

\item{x}{- The parameter in the X-axis: the time (as.Date object)}

\item{y}{-  The parameter in the Y-axis: the country names (character)}

\item{size}{- The parameter that represents the magnitude of the earthquake, indicating the size of the circles (=earthquake event): Richter scale value (numeric)}

\item{fill}{- The parameter that represents the deaths resulted from the earthquake, colouring the circles (=earthquake event)}
}
\value{
A geom layer with timeline, each timeline shows a country with its earthquake events in circles, the magnitude of an earthquake (size of
        circles) and the number of deaths (the colouring of the circles).

#@usage
#\\method{
#ggplot() +
#   geom_timeline(data, aes(x = DATE, y = COUNTRY, size = richterScaleValue, fill = DEATHS)) +
#   theme(legend.position = "bottom")
#   }
}
\description{
The geom_timeline visualises earthquakes events over time and per country in magnitude and number of deaths.
}
\details{
The geom_timeline function works together with ggplot2. This function visualises the earthquakes events over time (x aesthetics)
         and per country (y aesthetics) in magnitude (size parameter) and number of deaths (fill parameter.
}
\note{
error - An error will appear when a parameter is not given.
}
\examples{
\dontrun{
#library(magrittr)
#data <- eq_clean_data(file) \%>\%
#   filter(COUNTRY == c("China", "Usa", "Japan"),
#          DATE >= "1999-01-01",
#          DATE <= "2012-12-31")
ggplot() +
  geom_timeline(eq_clean_data(file) \%>\%
       filter(COUNTRY == c("China", "Usa", "Japan"),
             DATE >= "1999-01-01",
             DATE <= "2012-12-31"),
             aes(x = DATE, y = COUNTRY, size = richterScaleValue, fill = DEATHS)) +
  theme(legend.position = "bottom")
}

}
\keyword{datasets}
