% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes_plotting.r
\name{geom_timeline}
\alias{geom_timeline}
\title{Creating the geom -> geom_timeline()}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Mapping argument to the ggplot layer.}

\item{data}{A dataframe containing the data to plot.}

\item{stat}{Stat argument to the ggplot layer.}

\item{position}{Position argument to the ggplot layer.}

\item{na.rm}{na.rm argument to the ggplot layer.}

\item{show.legend}{show.legend argument to the ggplot layer.}

\item{inherit.aes}{inherit.aes argument to the ggplot layer.}

\item{...}{Extra Params.}
}
\value{
This function renders the plot into the current graphics device.
}
\description{
This function plots the dates of Earthquakes on a straight line.
The size of each point is directly proportional to the Ritcher scale value and the
colour is coded by the total deaths of each earthquake.
}
\examples{
\dontrun{
readr::read_delim("earthquakes-2021-09-27_15-28-44_+0200.tsv") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY \%in\% c("TURKEY") &
(lubridate::year(DATE) >= 2000 & lubridate::year(DATE) <= 2015)) \%>\%
ggplot() + geom_timeline( aes(x = DATE, size = MAGNITUDE, fill = DEATHS)) +
ggplot2::labs(size = "Richter scale value", fill = "# Deaths") + theme_time()
}

}
