% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\value{
This function returns a plot.
}
\description{
This geom plots a timeline of hurricanes in a country. The only thing you really
need to input is the \code{x} coordinate, which should be a date. You can optionally
provide a \code{y} coordinate, e.g. a country, to have more timelines stacked
vertically. The usual aes for geom_point apply.
}
