% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Earthquakes timeline}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings}

\item{data}{The dataset}

\item{stat}{stat layer}

\item{position}{Position adjustment}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{show.legend}{legend to display}

\item{inherit.aes}{false: default aesthetics}

\item{...}{Other arguments passed on to [layer()].}
}
\value{
time line plot of selected earthquakes.
}
\description{
Use for plotting a time line of earthquakes ranging from xmin to xmaxdates with a point for each earthquake.
}
\details{
Use \code{earthquakes} dataframe from this package or download new dataset from NOAA site.
}
\examples{
\donttest{
library(ggplot2)
a<-ggplot(earthquakes,aes(date,location_name,col=total_deaths,size=mag))+
geom_timeline()
}
}
