% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Make a time line label geom}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{ggplot parameter}

\item{data}{Earthquake data in the dataframe in format of NOAA earthquakes database (Dataframe), mandatory.}

\item{stat}{ggplot parameter}

\item{position}{ggplot parameter}

\item{na.rm}{ggplot parameter}

\item{show.legend}{ggplot parameter}

\item{inherit.aes}{ggplot parameter}

\item{...}{other ggplot parameter}

\item{xmindate}{aes parameters: Minimal date of the considering time frame (Date), mandatory.}

\item{xmaxdate}{aes parameters: Maximum date of the considering time frame (Date), mandatory.}

\item{event_date}{aes parameters: date of earthquake, usually name of the column in data (Date), mandatory.}

\item{richter}{aes parameters: Richter`s magnitude of earthquakes, usually name of the column in data (numeric), mandatory.}

\item{labels}{aes parameters: text labels of earthquakes, usually name of the column in data (character), mandatory.}
}
\value{
Timeline labels geom (ggplot2 object)
}
\description{
geom_timeline_label() function makes a ggplot2::Geom intended for adding annotations to the earthquake data.
This geom adds a vertical line to each data point with a text annotation (e.g. the location of the earthquake)
attached to each line. There is an option to subset to n_max number of earthquakes, where we take the n_max largest
(by magnitude) earthquakes. Aesthetics are x, which is the date of the earthquake and label which takes the column
name from which annotations will be obtained.
}
\examples{
  ggplot2::ggplot() +
  theme_void() +
  geom_timeline(data = earthquakes, aes(event_date = DATE,
                                  xmindate = 2000,
                                  xmaxdate = 2010,
                                  richter = EQ_PRIMARY,
                                  death = DEATHS,
                                  transparency = 0.3,
                                  colour = "red",
                                  scale = 0.5,
                                  layers = COUNTRY))+
  theme(legend.position="none") +
  geom_timeline_label(data = earthquakes, aes(event_date = DATE,
                                        xmindate = 2000,
                                        xmaxdate = 2010,
                                        richter = EQ_PRIMARY,
                                        layers = COUNTRY,
                                        labels = LOCATION_NAME))

}
