% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{Geom_Timeline_Label}
\alias{Geom_Timeline_Label}
\title{Labelling the timeline(s) of the earthquakes}
\format{An object of class \code{Geom_Timeline_Label} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
ggplot() +
  geom_timeline(data, aes(x = DATE, y = COUNTRY, size = richterScaleValue, fill = DEATHS)) +
  theme(legend.position = "bottom") +
  geom_timeline_label(data, aes(x=DATE, label = locations, n_maxVar = EQ_PRIMARY, n_max=10)) +
  labs(x = "Date", y = "Country", fill = "# deaths", size = "Richter scale value")
}
\arguments{
\item{x}{- The time parameter (as.Date object)}

\item{label}{- The location names parameter (character)}

\item{n_maxVar}{- The parameter that represents the magnitude of the earthquakes. This works together with the n_max parameter}

\item{n_max}{- Earthquakes beneath this value of magnitude will be shown on the map. The parameter of the magnitude needs to be
given in n_maxvar parameter.}
}
\value{
A geom visualising the locations of each earthquake event, indicated by a line.
}
\description{
The 'geom_timeline_label' function labels the timelines generated by the 'geom_timeline' function.
}
\details{
The 'geom_time_label' function works with ggplot2 and the 'geom_timeline' function. This function visualises the locations of the
         earthquakes, indicated by a line above each earthquake event.
}
\note{
warning - A warning appears when the no data is available to plot.
}
\examples{
\dontrun{
library(magrittr)
data <- eq_clean_data(file) \%>\%
  filter(COUNTRY == c("China", "Usa", "Japan"),
         DATE >= "1999-01-01",
         DATE <= "2012-12-31")
ggplot() +
  geom_timeline(data, aes(x = DATE, y = COUNTRY, size = richterScaleValue, fill = DEATHS)) +
  theme(legend.position = "bottom") +
  geom_timeline_label(data, aes(x=DATE, label = locations, n_maxVar = EQ_PRIMARY, n_max=10)) +
  labs(x = "Date", y = "Country", fill = "# deaths", size = "Richter scale value")
}

}
\keyword{datasets}
