% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Labelling the timeline(s) of the earthquakes}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{x}{- The time parameter (as.Date object)}

\item{label}{- The location names parameter (character)}

\item{n_maxVar}{- The parameter that represents the magnitude of the earthquakes. This works together with the n_max parameter}

\item{n_max}{- Earthquakes beneath this value of magnitude will be shown on the map. The parameter of the magnitude needs to be
given in n_maxvar parameter.}
}
\value{
A geom visualising the locations of each earthquake event, indicated by a line.
}
\description{
The 'geom_timeline_label' function labels the timelines generated by the 'geom_timeline' function.
}
\details{
The 'geom_time_label' function works with ggplot2 and the 'geom_timeline' function. This function visualises the locations of the
         earthquakes, indicated by a line above each earthquake event.
}
\note{
warning - A warning appears when the no data is available to plot.
}
\examples{
\dontrun{
library(magrittr)
data <- eq_clean_data(file) \%>\%
  filter(COUNTRY == c("China", "Usa", "Japan"),
         DATE >= "1999-01-01",
         DATE <= "2012-12-31")
ggplot() +
  geom_timeline(data, aes(x = DATE, y = COUNTRY, size = richterScaleValue, fill = DEATHS)) +
  theme(legend.position = "bottom") +
  geom_timeline_label(data, aes(x=DATE, label = locations, n_maxVar = EQ_PRIMARY, n_max=10)) +
  labs(x = "Date", y = "Country", fill = "# deaths", size = "Richter scale value")
}

}
