% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\value{
This function returns a plot.
}
\description{
This geom plots label over a timeline of hurricanes in a country. You need to input
the \code{x} coordinate, which should be a date, and a \code{label} parameter giving
the column from which you want to fish the labels. You can optionally
provide a \code{y} coordinate, e.g. a country, to have more timelines stacked
vertically. The usual aes for geom_point apply.
}
