% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline: Internal ggproto object for custom timeline geom}
\format{
An object of class \code{ggproto} defining the custom timeline geom.
}
\usage{
GeomTimeline
}
\description{
This is an internal ggproto object that defines the custom timeline geom for significant earthquakes. It should not be called directly, but instead used via the \code{geom_timeline} function.
}
\keyword{internal}
