% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\docType{data}
\name{GeomTimeLine}
\alias{GeomTimeLine}
\title{Custom Geom to plot earthquakes by date and magnitude.}
\format{An object of class \code{GeomTimeLine} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}) of length 7.}
\usage{
GeomTimeLine
}
\value{
Returns ggplot2 scatterplot
}
\description{
This Geom uses ggplot2, grid packages.
}
\keyword{datasets}

