% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Clean dataset}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\arguments{
\item{x}{Dates (date format)}

\item{y}{Country (a factor)}
}
\value{
This function returns a (hopefully cleaner) dataset

This function returns a Geom.
}
\description{
This function extracts locations from the hellspawn column LOCATION_NAME using
a deadly regex, and various trimming and beauty treatments from strings.
}
\examples{
\dontrun{
eq_location_clean(messy_data)
}

}
\keyword{datasets}
