% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geomtimeline.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{geom_timeline_label}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimelineLabel
}
\arguments{
\item{stat}{argument for calling ggplot2::layer, defaults to 'identity', a statistical transformation to be applied to the data}

\item{na.rm}{whether to remove NAs from the data.  Defaults to FALSE}

\item{x}{the dates for the timeline}

\item{y}{factor for timeline, each y value creates seperate timeline.  Can cause errors if not at least set to 0 in aes}

\item{labels}{how to label different quakes.  Defaults to ""}

\item{size}{how large the quakes are.  Used with n_max to determine which quakes to label first}

\item{n_size}{how many quakes to label.  Largest quakes are labeled first-otherwise random}
}
\value{
a graphical value containing labels and line-segments
}
\description{
creates labels for geom_timeline.  Goes with ggproto GeomTimelineLabel
}
\keyword{datasets}
