% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{A ggplot2 geom object for creating NOAA earthquake timeline labels.}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.}
\usage{
GeomTimelineLabel
}
\description{
A ggplot2 geom object for creating NOAA earthquake timeline labels.
}
\keyword{datasets}
