% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel: Internal ggproto object for custom timeline label geom}
\format{
An object of class \code{ggproto} defining the custom timeline label geom.
}
\usage{
GeomTimelineLabel
}
\description{
This is an internal ggproto object that defines the custom timeline label geom for significant earthquakes. It should not be called directly, but instead used via the \code{geom_timeline_label} function.
}
\keyword{internal}
