% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\docType{data}
\name{GeomTimeLineLabel}
\alias{GeomTimeLineLabel}
\title{Custom Geom to plot earthquake labels for date or location, magnitude, deaths.}
\format{An object of class \code{GeomTimeLineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeLineLabel
}
\value{
Returns ggplot2 labels
}
\description{
This Geom uses ggplot2, grid packages.
}
\keyword{datasets}

