% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes_plotting.r
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Creating a Geom -> GeomTimelineLabel}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimelineLabel
}
\arguments{
\item{data}{A dataframe with the data to plot.}

\item{x}{Data to be plotted on the x-axis.}

\item{y}{Data to be plotted on the y-coordinate.}

\item{size}{Column of data proportional to the size of each point.}

\item{colour}{Column of data depicting the colour of each point.}

\item{shape}{Column of data in charge of the shape of each of the points.}

\item{alpha}{Column of data used to apply the alpha. \cr \cr}

\item{fill}{String introduced to modify the fill of the points.S}
}
\value{
The function returns a Geom to be used by geom_timeline to render
the images into the plot.
}
\description{
This Geom returns a ggproto object for our geom to inherit from. It also includes
labels for each of the earthquakes plotted. The creation of this function is documented
inside the function for easy comprehension and to reduce the size of this description.
}
\keyword{datasets}
