% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes_plotting.r
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Creating a Geom -> GeomTimelineLabel}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimelineLabel
}
\value{
The function returns a Geom to be used by geom_timeline to render
the images into the plot.
}
\description{
This Geom returns a ggproto object for our geom to inherit from. It also includes
labels for each of the earthquakes plotted. The creation of this function is documented
inside the function for easy comprehension and to reduce the size of this description.
}
\keyword{datasets}
