% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw}
\alias{raw}
\title{Significant Earthquakes Database}
\format{A data frame with columns:
\describe{
 \item{COUNTRY}{}
 \item{DAMAGE_DESCRIPTION}{}
 \item{DAMAGE_MILLIONS_DOLLARS}{}
 \item{DAY}{}
 \item{DEATHS}{}
 \item{DEATHS_DESCRIPTION}{}
 \item{EQ_MAG_MB}{}
 \item{EQ_MAG_MFA}{}
 \item{EQ_MAG_ML}{}
 \item{EQ_MAG_MS}{}
 \item{EQ_MAG_MW}{}
 \item{EQ_MAG_UNK}{}
 \item{EQ_PRIMARY}{}
 \item{FLAG_TSUNAMI}{When a tsunami was generated by an earthquake, "Tsu" appears in the Assoc Tsu column. There should also be a "TSU" link in the Assoc column which is linked to the tsunami event database. The TSU link will display additional tsunami event information in a new browser window.}
 \item{FOCAL_DEPTH}{}
 \item{HOUR}{}
 \item{HOUSES_DAMAGED}{}
 \item{HOUSES_DAMAGED_DESCRIPTION}{}
 \item{HOUSES_DESTROYED}{}
 \item{HOUSES_DESTROYED_DESCRIPTION}{}
 \item{INJURIES}{}
 \item{INJURIES_DESCRIPTION}{}
 \item{INTENSITY}{}
 \item{I_D}{The unique numeric identifier of the record.}
 \item{LATITUDE}{}
 \item{LOCATION_NAME}{}
 \item{LONGITUDE}{}
 \item{MINUTE}{}
 \item{MISSING}{}
 \item{MISSING_DESCRIPTION}{}
 \item{MONTH}{}
 \item{REGION_CODE}{}
 \item{SECOND}{}
 \item{STATE}{}
 \item{TOTAL_DAMAGE_DESCRIPTION}{}
 \item{TOTAL_DAMAGE_MILLIONS_DOLLARS}{}
 \item{TOTAL_DEATHS}{}
 \item{TOTAL_DEATHS_DESCRIPTION}{}
 \item{TOTAL_HOUSES_DAMAGED}{}
 \item{TOTAL_HOUSES_DAMAGED_DESCRIPTION}{}
 \item{TOTAL_HOUSES_DESTROYED}{}
 \item{TOTAL_HOUSES_DESTROYED_DESCRIPTION}{}
 \item{TOTAL_INJURIES}{}
 \item{TOTAL_INJURIES_DESCRIPTION}{}
 \item{TOTAL_MISSING}{}
 \item{TOTAL_MISSING_DESCRIPTION}{}
 \item{YEAR}{Century and Year}
}}
\source{
NOAA (National Centers for Environmental Information) \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
raw
}
\description{
Significant Earthquakes Database
}
\examples{
\dontrun{
 raw
}
}
\keyword{datasets}
