% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{tremors}
\alias{tremors}
\title{tremors}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6161 rows and 48 columns.}
\usage{
tremors
}
\description{
tremors
}
\note{
located in /data/tremors.rda
}
\references{
NOAA
}
\author{
Christopher Pearson
}
\keyword{data}
